This reverts commit 71d71b0615d83fbca0391e6a901eb2d9cbab10aa.

+++ Shopezly/src/main/java/com/shopezly/repository/ProductRepo.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.shopezly.model.Product;
 
 @Repository
-public interface ProductRepo extends JpaRepository<Product, Integer> {
+public interface ProductRepo extends JpaRepository<Product, Long> {
     
 }
